analytics: Split test_views.py.
This will aligns the test organization with an upcoming reorganization
of the code itself.
